Add the apt-dev pubkey when setting up acceptance tests
This fixes an issue where, when the same version of the package is
the latest in both production and dev repos, apt appears to use the metadata
from the dev repo, but download the package from the production repo
(because it doesn't have the dev key). Because the production version is
signed, the package doesn't match the metadata.